+++ src/main/java/net/grid/vampiresdelight/client/event/ToolTipEvents.java
@@ -7,6 +7,7 @@ import de.teamlapen.vampirism.items.GarlicBreadItem;
 import de.teamlapen.vampirism.items.VampirismItemBloodFoodItem;
 import net.grid.vampiresdelight.VampiresDelight;
 import net.grid.vampiresdelight.common.VDConfiguration;
+import net.grid.vampiresdelight.common.item.FactionConsumableItem;
 import net.grid.vampiresdelight.common.tag.VDTags;
 import net.grid.vampiresdelight.common.utility.VDHelper;
 import net.grid.vampiresdelight.common.utility.VDIntegrationUtils;
@@ -36,6 +37,10 @@ public class ToolTipEvents {
         Player player = VampirismMod.proxy.getClientPlayer();
 
         if (!VDConfiguration.COLORED_TOOLTIPS.get() || stack.isEmpty()) {
+            return;
+        }
+
+        if (item instanceof FactionConsumableItem factionConsumableItem && !factionConsumableItem.hasColoredTooltipMargins(stack, player)) {
             return;
         }